+++ src/main/kotlin/com/chrzanekk/simulatorpandemic/service/dto/SimulationDTO.kt
@@ -1,6 +1,7 @@
 package com.chrzanekk.simulatorpandemic.service.dto
 
 import com.chrzanekk.simulatorpandemic.domain.SimulatedPopulation
+import com.sun.jdi.IntegerValue
 import java.math.BigDecimal
 
 data class SimulationDTO(val id: Long,
@@ -9,7 +10,7 @@ data class SimulationDTO(val id: Long,
                          val initialNoOfInfectedPpl: BigDecimal,
                          val infectionRate: BigDecimal,
                          val mortalityRate: BigDecimal,
-                         val infectedToCureDuration: BigDecimal,
-                         val infectedToDeathDuration: BigDecimal,
-                         val simulationDays: BigDecimal,
+                         val infectedToCureDuration: Int,
+                         val infectedToDeathDuration: Int,
+                         val simulationDays: Int,
                          val simulatedPopulation: List<SimulatedPopulationDTO>)
